// psk-ui/src/app/services/models/test-template-update-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TestTemplateCommentRequest } from '../models/test-template-comment-request';
export interface TestTemplateUpdateRequest {
  comments?: Array<TestTemplateCommentRequest>;
  imageUrl?: string;
  isActive?: boolean;
  scoringStrategy?: 'SIMPLE_LINEAR' | 'WEIGHTED' | 'PERCENTAGE';
  subTitle?: string;
  title: string;
}